refactor: moved DTOs at right place
